refactor into OOP library
